requirement.txt
torch
torchvision
numpy
opencv-python
timm
einops
basicsr
thop
huggingface_hub
pywavelet
clip
open_clip
collections
functools
lpips
scikit-image
pillow
hdf5storage
ninja
lmdb
requests
